+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/domain/post/Content.java
@@ -1,0 +1,5 @@
+package wooteco.team.ittabi.legenoaroundhere.domain.post;
+
+public class Content {
+
+}
