+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/repository/CommentRepository.java
@@ -3,8 +3,8 @@ package wooteco.team.ittabi.legenoaroundhere.repository;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.team.ittabi.legenoaroundhere.domain.Comment;
-import wooteco.team.ittabi.legenoaroundhere.domain.State;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.Comment;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.State;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
