+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/repository/ImageRepository.java
@@ -1,7 +1,7 @@
 package wooteco.team.ittabi.legenoaroundhere.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.team.ittabi.legenoaroundhere.domain.Image;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.image.Image;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
