+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/repository/LikeRepository.java
@@ -1,7 +1,7 @@
 package wooteco.team.ittabi.legenoaroundhere.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.team.ittabi.legenoaroundhere.domain.Like;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.like.Like;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
