+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/repository/PostRepository.java
@@ -4,8 +4,8 @@ import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.team.ittabi.legenoaroundhere.domain.Post;
-import wooteco.team.ittabi.legenoaroundhere.domain.State;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.Post;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.State;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
