+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/service/LikeService.java
@@ -3,9 +3,9 @@ package wooteco.team.ittabi.legenoaroundhere.service;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.team.ittabi.legenoaroundhere.domain.Like;
-import wooteco.team.ittabi.legenoaroundhere.domain.Post;
-import wooteco.team.ittabi.legenoaroundhere.domain.State;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.Post;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.State;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.like.Like;
 import wooteco.team.ittabi.legenoaroundhere.domain.user.User;
 import wooteco.team.ittabi.legenoaroundhere.dto.LikeResponse;
 import wooteco.team.ittabi.legenoaroundhere.exception.NotExistsException;
